+++ src/events/DefineCustomFilterOptionsEvent.php
@@ -2,9 +2,14 @@
 
 namespace wsydney76\contentoverview\events;
 
+use craft\elements\User;
+use Illuminate\Support\Collection;
+use wsydney76\contentoverview\models\Filter;
 use yii\base\Event;
 
 class DefineCustomFilterOptionsEvent extends Event
 {
-    public array $filter;
+    public User $user;
+    public Filter $filter;
+    public Collection $options;
 }